backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "google-genai>=0.6.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.45.0",
    "yfinance>=0.2.50",
]
